feat: add types to drag-and-drop component
